// src/app/shared/enums/auction-phase-text.enum.ts
export enum AuctionPhaseTextEnum {
  auctionCancelled = 'auctionCancelled',
  futureAuction = 'futureAuction',
  upcomingAuction = 'upcomingAuction',
  liveAuction = 'liveAuction',
  auctionClosed = 'auctionClosed',
  pickupComplete = 'pickupComplete',
  auctionComplete = 'auctionComplete',
  pastAuction = 'pastAuction',
}